requirements.txt
torch>=1.9.0
timm>=0.6.0
numpy>=1.21.0
Pillow>=8.3.0
matplotlib>=3.4.0
tqdm>=4.62.0
editdistance
onnxscript